+++ NEWS.md
@@ -8,8 +8,9 @@ output: html_document
 Details of changes made (**/**/2022):
 
 1. add 'stm' function
-1. Modify 'psim_artif' & 'psim_real' msg "Preview the final spatial and temporal model?"? Yes or No "Continue?" `vignette` to guide the use of the package
-Vignette modified accordingly. interpolation..
+2. Modify 'psim_artif' & 'psim_real' to allow preview of expected spatial and temporal model?"? 
+to guide the use of the package
+3. Editted the vignette accordingly. interpolation..
 2. Implement network space (for the 'planar' and 'network'.... 'real' point snapped to nearest segment)
 3. Address the computational time for the 'psim_real'